Trying to merge offline stuff from Saria/Sirius

+++ trunk/basonejlib/src/com/baseoneonline/java/tools/ImprovedNoise.java
@@ -2,14 +2,14 @@ package com.baseoneonline.java.tools;
 
 
 	public final class ImprovedNoise {
-		   static public float noise(float x, float y, float z) {
+		   static public double noise(double x, double y, double z) {
 		      int X = (int)Math.floor(x) & 255,                  // FIND UNIT CUBE THAT
 		          Y = (int)Math.floor(y) & 255,                  // CONTAINS POINT.
 		          Z = (int)Math.floor(z) & 255;
 		      x -= Math.floor(x);                                // FIND RELATIVE X,Y,Z
 		      y -= Math.floor(y);                                // OF POINT IN CUBE.
 		      z -= Math.floor(z);
-		      float u = fade(x),                                // COMPUTE FADE CURVES
+		      double u = fade(x),                                // COMPUTE FADE CURVES
 		             v = fade(y),                                // FOR EACH OF X,Y,Z.
 		             w = fade(z);
 		      int A = p[X  ]+Y, AA = p[A]+Z, AB = p[A+1]+Z,      // HASH COORDINATES OF
@@ -24,11 +24,11 @@ package com.baseoneonline.java.tools;
 		                             lerp(u, grad(p[AB+1], x  , y-1, z-1 ),
 		                                     grad(p[BB+1], x-1, y-1, z-1 ))));
 		   }
-		   static float fade(float t) { return t * t * t * (t * (t * 6 - 15) + 10); }
-		   static float lerp(float t, float a, float b) { return a + t * (b - a); }
-		   static float grad(int hash, float x, float y, float z) {
+		   static double fade(double t) { return t * t * t * (t * (t * 6 - 15) + 10); }
+		   static double lerp(double t, double a, double b) { return a + t * (b - a); }
+		   static double grad(int hash, double x, double y, double z) {
 		      int h = hash & 15;                      // CONVERT LO 4 BITS OF HASH CODE
-		      float u = h<8 ? x : y,                 // INTO 12 GRADIENT DIRECTIONS.
+		      double u = h<8 ? x : y,                 // INTO 12 GRADIENT DIRECTIONS.
 		             v = h<4 ? y : h==12||h==14 ? x : z;
 		      return ((h&1) == 0 ? u : -u) + ((h&2) == 0 ? v : -v);
 		   }
